Eliminate credentials from the image
XXX: Initializations of `protonvpn` are done on container launching,
not on image building.
This is because `protonvpn` saves credentials in plain text format, but
consider initializing on build and prune just credentials.
To achieve this, we should investigate what `protonvpn init` generates
indeed. Is it safe to remove just credentials and retain the others?